requirements-test.txt
-r requirements.txt
pytest==5.1.2
pytest-django==3.5.1
ipython==7.8.0
jupyter==1.0.0
pylint==2.3.1
pylint-django==2.0.11